Database, dbcontext.

+++ WebImage/Context/FileContent.cs
@@ -15,8 +15,8 @@ namespace WebImage.Context
         public string Name { get; set; }
         public string Title { get; set; }
         public string Extension { get; set; }
-        public float LengthKB { get; set; }
-        public float LengthMB { get; set; }
+        public double LengthKB { get; set; }
+        public double LengthMB { get; set; }
         public string Url { get; set; }
         public string Category { get; set; }
         public byte[] Content { get; set; }
